New features: 1) A UDL receives a JSON configuration object given by 'user_defined_logic_config_list' in dfgs.json; 2) DLL interface: get_observer(ctxt,json_udl_config); 3) UserDefinedLogic::get_observer(ctxt,json_udl_config); 4) UserDefineLogicManager::get_observer(udl_id,json_udl_config); 5) smoke test case for pipeline_udl;
TODO: There is an issue with the cascade service client. If a client is a Cascade member, it will use the group_ptr instead of the extercal client to contact a node when needed. In such a case, it will call group_ptr->get_nonmember_subgroup<>() to get the handle to the target subgroup. However, it might be a member of that subgroup and throw exception on get_nonmember_subgroup<>(). To fix this issue, it needs to test if itself is in the target subgroup or not, by group_ptr->get_my_shard<>(). This has been fixed for trigger_put. Need to fix it for all other CascadeServiceClient APIs.