Initial work on SHA UDL and SignatureStore
This doesn't actually work, but it compiles, and it's enough of a
sketch of my ideas to share. The ShaHashObserver is meant to be a UDL
installed on each PersistentCascadeStore node of the CascadeChain, and
it will send the hash it computes to a "parallel" SignatureCascadeStore
node. SignatureCascadeStore will automatically sign each delta for its
key-value store of hashes, but it needs a way to retrieve those
signatures, and I haven't figured that out yet.