Fixed CMakeLists to locate Derecho in non-system paths
In the existing setup, it was not possible to install Derecho anywhere
other than /usr/local (or another "system" path) even if find_package
successfully found Derecho in that non-standard location. This is
because we didn't actually do anything with the result of find_package,
which is either some variables named derecho_INCLUDE_DIRS and
derecho_LIBRARIES, or an IMPORT target named derecho::derecho.

This works, although it still has to use derecho_INCLUDE_DIRS even
though the import target derecho::derecho is successfully loaded. I
think it's because each subdirectory of cascade does not inherit the
parent's dependencies (listed in target_link_libraries(cascade...)) but
I can debug that more later. At least now I can run Cascade with a
"user" copy of Derecho installed in my home directory.