Fixed CMakeLists to better declare dependencies on Readline and rpclib
Readline is required, but wasn't declared as a required package, so if
you don't have it installed you wouldn't get an error until client.cpp
failed to compile. Also, rpclib provides an import target named
rpclib::rpc, so we should use that instead of the plain library name in
target_link_libraries.