Made SignatureStore part of Service, and revised sha_hash_udl
After talking with Weijia, the new plan is to just add
SignatureCascadeStore to the default Service templates, so every Cascade
service has the ability to add signatures, and then rely on
configuration files to make it act like CascadeChain (no Volatile store,
no Trigger store, every update submitted to PersistentStore gets hashed
and set to the SignatureStore). This means the SHA-hashing UDL can use
the object pool name in the outputs map to identify the SignatureStore,
as long as the object pools were set up correctly.