Added an object-pool version of get_signature
This was harder than it looks because it's invalid to call get_signature
on a subgroup that is not a SignatureCascadeStore, so I have to prevent
the compiler from even attempting to generate a call to
get_signature<PersistentCascadeStore>; the type_recursive_get functions
don't have to worry about this because get<FirstType> is always valid to
compile regardless of whether FirstType is the desired type (which is
chosen at runtime based on the value of type_index).

Also, fixed a dumb mistake in sha_hash_udl and replaced some runtime
type-checks with compile-time static_asserts.