Added a client for testing CascadeChain, and some more configs
To make things easier, I wrote a client that copies some of client.cpp
but only has the object-pool versions of the commands, then added the
new commands I need for proper client-side behavior on CascadeChain. We
might want to copy these commands (like put_with_signature) back to the
"main" client at some point, if we want to stick with the design of
making the Signature subgroup always a part of the Cascade Service.