Added a field to ObjectWithStringKey and continued work on client
I'm not sure if this is the best solution, but it's really hard to keep
track of the link between object versions in PersistentCascadeStore and
hash/signature versions in SignatureCascadeStore. Adding another header
field in ObjectWithStringKey allows a client to retrieve the correct
hash and signature for a given object version.

I also needed to add another P2P RPC call to SignatureCascadeStore to
retrieve the "previous version" when it is unrelated to the desired key.