Reverted changes to Object and added a version map to SignatureStore
After a discussion with Weijia, I realized it would be much better to
store the mapping from data-object versions to hash-object versions as a
field within SignatureCascadeStore, instead of in another header in
ObjectWithStringKey. SignatureCascadeStore can translate the desired
data object version to a hash object version itself, rather than
expecting clients to know which hash object version to request.