Added notification support to ServiceClient
There aren't (yet) any notifications that a client can receive, but I
added NotificationSupport to SignatureCacadeStore and mocked up some
notification-subscribing methods to make sure ServiceClient would
compile when given a Cascade service that has notifications. The new
ServiceClient requires the Main method to provide no-argument factory
functions for each Cascade type in order to enable notifications, and
exposes the register_notification function from ExternalClientCaller.

Since the no-argument factories will only be used to construct
non-functional local copies of the Cascade types, they can construct
PersistentCascadeStore (and the similar SignatureCascadeStore) with a
null PersistentRegistry, and this shouldn't cause a problem as long as
the client's PersistentCascadeStore doesn't try to actually persist
anything. This is the approach I used in chain_only_client.cpp.