Split cascade.hpp into multiple files
This is a somewhat major reorganization to make it easier to work on
different branches. Instead of all the CascadeStore types being defined
and implemented in one pair of files (cascade/cascade_impl.hpp), each
class is now in its own pair of files. cascade.hpp is now a meta-file
that includes the four component files, so clients can still just
include <cascade/cacade.hpp> to use Cascade. I also split
src/service/server.cpp into a header and source file to make it easier
to reuse the CDPO definition, which I needed to do in my cascade_chain
branch.

There are no code changes in this commit, and everything still compiles.