Fixed SignatureStore's use of "stable" parameter
After a discussion with Weijia I figured out how to accommodate the new
API option "stable" in a get() request. The SignatureStore should always
require a version number, not CURRENT_VERSION, and it will wait for the
translated version number's stability if stable is requested. I also
updated client.cpp and chain_client.cpp to use the new API.