Added PersistenceObserver and began implementing signature notifications
The PersistenceObserver class is a general-purpose helper that observes
persistence notifications coming from a Derecho group and lets other
classes register actions that should fire when a particular persistence
event occurs. It processes these actions in a new thread, separate from
the Derecho thread that delivers the persistence callbacks, to avoid
slowing down Derecho. I added a PersistenceObserver to the
CascadeContext, and made it retrievable from the type-erased
ICascadeContext interface, since this does not actually depend on the
CascadeTypes template parameters. This allowed me to implement
SignatureCascadeStore's request_notification and
subscribe_to_notifications methods as actions registered with the
PersistenceObserver.

Also updated the CMake minimum version to 3.16, since that is the
current "stable" version in Ubuntu 20.04 LTS and I was having trouble
getting the old version of CMake to resolve the Derecho dependency
properly.