Replaced nullptr casts with declval to get rid of warnings
Later versions of GCC produce copious warnings about using a null 'this'
pointer to call a member function, even when that null pointer is within
a decltype context and thus will never be executed. Using std::declval
is the "proper" way to get an object and call a member function inside a
decltype expression, instead of casting nullptr to the desired object
type.