Added debugging statements, made client subscribe to notifications
The client still isn't getting the notifications it requests, and I'm
trying to figure out why. Part of the problem is that requesting a
single notification for a specific version creates a race condition: the
request for notification can arrive at the SignatureStore before the
put_and_forget() from the SignatureStore does, and thus the
SignatureStore won't know about that version yet.