Created a context object for the chain client
Refactored the chain_only_client code to store all of the client's state
in a ChainClientContext object rather than in static global variables.
This meant changing the "command" API to pass around this object rather
than a ServiceClientAPI. Also fixed some bugs in the signature-related
commands, such as get_and_verify attempting to call SignatureStore's
get() with CURRENT_VERSION if the user doesn't specify a version.