Started creating a WanAgent integration
Right now the plan is for PersistentStore and SignatureStore to load a
UDL that runs WanAgent and uses it to forward any data it receives
(actual data or a hash/signature) to the backup sites. For the
SignatureStore's UDL to get access to the signature, SignatureStore
needs to initiate a trigger_put to itself once the signature is
generated, and add the signature to the body of the object, because only
the object is visible to the UDL (not metadata like the signature).