Improved config files, worked on the wan_backup_udl
Filled in the wanagent.json files that come with each node of
chain_pipeline_cfg to have the proper IP addresses for a Fractus
deployment, since that's probably where I'll test the backup site. Fixed
the derecho.cfg for nodes 5 and 6 to use a different leader so they are
a separate group from nodes 0-4. Fixed dfgs.json for the primary site to
actually load the WAN Backup UDL, and created an empty dfgs.json for the
backup site so Cascade would stop complaining about the missing
dfgs.json (even though there are no UDLs for the backup site).

Also, started implementing a way for WAN Backup UDL to send
notifications when it detects that a backup message has been
acknowledged, but got stuck.