Fixed one more compile error
For some reason, I was able to get everything to compile in the last
commit without fixing this error in cascade_client_py.cpp, but it showed
up when I recompiled on a different computer. It's another merge problem
caused by the fact that I changed the return type of put() and remove()
in my branch.