Moved WanAgent to run inside SignatureStore instead of as a UDL
This is a mostly complete implementation, but hasn't been tested yet,
because it exposes a bug in mutils-serialization: The support for
std::tuple and std::map is incomplete. As a result, the new RPC
functions in SignatureCascadeStore that have tuple and map arguments
won't compile. Thus, I have to stop here and go work on fixing
mutils-serialization before I can continue.