Fixed a minor error in config files
In that last commit, I accidentally committed an "old" version of two of
the files in which node 4 was the leader of the backup site, even though
the rest of the files were the "new" version where node 5 is the leader
of the backup site. Also, CMakeLists.txt needs another ln -sf command
for the new n7 config folder.