Added configuration option for primary/backup server role
Now that SignatureCascadeStore runs WanAgent, it won't work for the
backup server to be a separate binary that runs WanAgent in the main
thread. Both primary and backup will run SignatureCascadeStore, which
means they will both run the same WanAgent code, so there needs to be a
configuration option that tells SignatureCascadeStore whether it behaves
as the primary or the backup.

Also removed the obsolete wan_backup_udl.