Changed version_tuple order to be compatible with master branch
My latest merge with master caused compile errors in the new C# client
because it expected version_tuple to be <version_t, uint64_t>, and this
made me realize it was probably a bad idea to declare my expanded
version_tuple with the new previous-version entries inserted in the
middle. Instead, my version_tuple should add the new entries at the end,
so the first two entries are still the version and the timestamp, and
code that expects a two-entry tuple with version and timestamp will
still work by just ignoring the two extra entries. This makes my change
more "backward compatible" and should reduce conflicts in future merges.