Sorted clang-format, added a few new options
For some reason our clang-format file was almost alphabetized, but not
quite, which made it hard to read. I also discovered a few new options
that clang-format has added since we last checked, and explicitly wrote
them down using the default values we are currently expecting, to guard
against future changes to the defaults. This change should not result in
any changes to our clang-format style.