Added message ID to notifications, created perftest configuration
The sha_hash_udl now propagates the optional message_id field of the
ObjectWithStringKey to the SignatureStore when ENABLE_EVALUATION is on.
SignatureCascadeStore's signature notifications propagate it back to the
client, prepended to the normal message body, so the client can create a
log entry for when a particular update (identified by its message ID)
has finished being signed.

Created a set of configuration files to help me manually test the
performance test setup without using a Python script. These include
private keys to minimize the hassle of generating and copying keys to
all the nodes, so obviously the committed private key should never be
used in production.

Also, fixed a warning by removing the "inline" declaration from a method
in service.hpp.