Made perf_put dump timestamps on signature object pool
I didn't realize that the PerfTestClient's perf_put function was
responsible for calling dump_timestamp on the servers after the
evaluation function (e.g. eval_put) finished. The signature-storage
servers weren't generating timestamp.log, because PerfTestClient only
sent the dump_timestamp command to the object pool specified as the
target of perf_put, which is the storage pool. Now it calls
dump_timestamp on both object pools when called in signatures mode.