Ensured perftest client won't run out of memory for large objects
The macro-defined constant NUMBER_OF_DISTINCT_OBJECTS meant the
make_workload function always created 4096 objects, but if payload_size
was 100MB, this would attempt to create 400GB of objects. As a quick
solution, I made make_workload take the number of objects as a
parameter, and then perftest computes it to be either 4096 or the
maximum number of objects that can fit in 16GB, whichever is smaller.