Added unsubscribe_from_notifications to SignatureStore
The perftest client subscribed to notifications at the beginning of each
run, but had no way of unsubscribing from them, which meant the server
accumulated an additional subscription request in each round and would
send the client more and more notifications each time. To make each run
of the perftest client idempotent, it needs to unsubscribe from
signature notifications at the end of the round.