Removed redundant CMakeLists declarations
According to what I can find on the Internet about how to structure
CMake projects, only the top-level CMakeLists.txt should have the
"minimum required" and "disable in-source builds" declarations at the
top, since these are supposed to be set once for the entire project. All
the subdirectory CMakeLists.txt files should be fragments that only
process that directory.

This is a manual cherry-pick of commit 316222aa4b9783532e749fb89dfbff3de9da9bf5
from the cascade_chain branch, which has diverged too much for an actual
cherry-pick to work.