Slight improvement to CMakeLists for the Python client
On some systems (i.e. newer Ubuntu) the Python executable is named
python3 by default, and "python" does not exist unless you install the
"python-is-python3" package. With CMake 3.15+, you can use the
FindPython module to find the exact name of the Python executable on the
system. Also, the Python client's CMakeLists should inform the user
whether Python support is enabled or disabled the same way the C#
client's CMakeLists does.