Another small change to perftest
The signature notifications don't have the correct version when I put
directly to the signature group (they have the fake version I set).
Fortunately the notifications also include the message ID, so I didn't
need to be tracking the last version anyway - I could just set
all_signed when the notification arrives for the last message ID, which
is easily determined as soon as the sending loop is finished.