Added a non-copying constructor for Blob
To reduce unnecessary memory copies when constructing Cascade
notification messages (which contain Blob objects), I added a
constructor for Blob that takes a unique_ptr to a buffer and takes
ownership of it. This allows me to create a buffer, put the message in
it, and then give the buffer to the Blob (move it) instead of creating a
Blob that copies the message into a new buffer. unique_ptr<uint8_t[]>
allows me to express ownership transfer, which is impossible to express
with a raw pointer to uint8_t.

While doing this, I realized that Blob was using the old C malloc and
free functions, which can't interoperate with the C++ new operator. I
changed it to use new uint8_t[] and delete[] so that it is safe to take
ownership of a unique_ptr<uint8_t[]>.