Created a perftest function to test a get-only workload
The perf_put functions say they accept a read/write ratio parameter, but
in fact they ignore it and always do puts, so we need a new function
that does a performance test for gets. We probably needed a new function
anyway for a get-only workload, since if the read/write ratio is 100%
reads there will be no objects in the object pool to read; a read-only
test function needs to put the test objects in the pool first, then
start the experiment.

This might not work yet, and isn't complete, because there's no way to
call perf_get from the main client.cpp yet.