Finished writing framework for perf_get test
The new eval_get test in PerfTestServer now has entry points for both
the object-pool and specific-shard versions in PerfTestClient, and
commands to call both of these versions in client.cpp.

Also fixed the test to use multi_get instead of get when the request is
for the current version, since apparently get() for the current version
does not (always) become a multicast any more.