Fixed bugs in eval_get
I somehow forgot to use objects.size() instead of the old macro, and it
still compiled because the macro was still defined. Also the calculation
to ensure the objects fit in the persistent log had an off-by-one error:
each object will have log_depth + 1 versions in the log, because the
current version (the 0th update) also counts.