Reduced num_distinct_objects again to handle larger objects
I ran into another problem when testing perf_get with long log_depth: If
the size of each object/update is too large, the servers can actually
run out of storage when trying to add a long log to all the objects.
This doesn't show up in the perf_put tests because they don't actually
put all of the possible objects in the workload, just a random subset of
them based on the value of the clock during the test.