Added a get_by_time perf test
Right now I only created the front-end boilerplate for the shard-based
API (not the object-pool API) since that's the only one we're testing.
Also, right now this assumes we want to remember the exact timestamp of
each object's desired version instead of requesting the same timestamp
for every get request.