Fixed bugs in eval_get_by_time
Now that I was able to test it on Fractus I discovered some bugs. Weijia
was right about the need to add a slight delta to each timestamp
received from put() to ensure you can get_by_time() successfully with
that timestamp: doing a get_by_time() with the exact timestamp sometimes
results in "No data found for key before version" because the object was
actually assigned a timestamp slightly later than the one returned by
put() and the get() becomes a request for a time before the object
existed.