Changed eval_get_by_time to always request the same object
Since it takes time to put all the objects in the store, the timestamps
for the oldest versions of each of those objects will be pretty far
apart; they definitely won't all be ms_in_past milliseconds in the past.
Requesting the same object every time seems too simple, but it's the
only way I can think of to make the get_by_time() requests all have a
similar log-searching requirement.