Rudimentary system for sending signature notifications from backup site
This is fairly stupid but it ought to work for the short term to get
some experiment data. Each node on the backup site listens to accept
exactly one connection from a wide-area client on the newly-configured
remote_client_port, and then sends that client a notification every time
a signature finishes persisting. The perftest client spawns a second
thread, in addition to the future-consuming thread, to connect to the
backup site's SignatureCascadeStore node on the remote_client_port and
read the notifications as they come in. The IP of the node to contact on
the backup site must be hand-written in the config file, not a
command-line parameter, in order to avoid changing the perftest client's
RPC interface.