Properly shut down remote-client thread
Although the simple thread I created to accept a connection from the
remote client was supposed to exit after accepting a single connection,
this only works on the backup-site replica that actually gets a
connection. On the other replica, the remote-client thread will prevent
the server from shutting down because it is stuck waiting for a
connection. Also added some debugging log statements.