Moved listening thread start to new-view callback
When state is transferred to a new node during View change, the new node
first calls the normal constructor, then the deserialization
constructor. Since the deserialization constructor isn't a "true" move
constructor, the old object still exists when the new one is being
deserialized, even though the new object is about to overwrite the old
one. This means attempting to start the thread in the deserialization
constructor will fail, because another copy of it is already bound to
the remote client port. On the other hand, the new-view callback is only
called after state transfer has finished and the old object has
definitely been destructed.