Increased completion queue timeout in config files
The sst_poll_cq_timeout is different from the heartbeat, and represents
the time the SST polling thread has to notice a completion from a
heartbeat message. When servers are heavily CPU-bound, the polling
thread might not run promptly enough to see the completion by a short
timeout, even if the heartbeat was successful and nothing has failed.