Updated Cascade to build with spdlog 1.12
The next version of Derecho will depend on spdlog 1.12 instead of
spdlog 1.9.2, and this library update introduced breaking changes. A few
new lines of code are needed to allow Cascade to compile against the
Derecho version that uses spdlog 1.12, and Cascade's own CMakeLists
should specify that it expects the newer spdlog version (since it will
no longer compile against spdlog 1.9.2 with these changes).