Created a DeltaMap for the version map in SignatureStore
Putting a std::map in Persistent<T> is very inefficient, since
Persistent<T> will save the entire map for every version even if only
one key-value entry has changed. It's much better to use Persistent's
Delta API, like DeltaStoreCore does, so each version in the persistent
log only contains the key-value pair that was added/removed in that
update. I've been meaning to do this for a while but finally got around
to it. The new DeltaMap is mostly just a copy of DeltaCascadeStoreCore
with no lockless-read support (not needed right now) and no IK template
parameter (never actually used in DeltaCascadeStoreCore and no reason to
use it in this class).