Fixed client code to verify signatures on delta log entries
Another side-effect of the change to DeltaCascadeStoreCore DeltaType
that I didn't notice in my first round of changes: The client code needs
to recreate an entire DeltaType, not just an ObjectWithStringKey, in
order to correctly verify the signature. This means DeltaType needs to
have to_bytes implemented in a way that matches currentDeltaToBytes.