Fixed log formatting of enums to be compatible with spdlog 1.15
The latest version of spdlog (1.15) bundles a newer version of fmtlib,
which removes its default formatting behavior for enum types. Enum
values must now either have a formatter specialization declared or be
wrapped in a call to fmt::underlying() to print them as integers. (The
previous behavior was to automatically print enum values as integers).