Added a log formatter for ShardMemberSelectionPolicy
This enum will need a formatter overload if we ever want to print out
its value in a logging statement. The main Cascade branch doesn't
currently do that, but the cascade_chain branch does (to debug member
selection policies in the SignatureStore subgroup), so I know it will be
needed. It seems better to fix it in the main branch and then inherit
the changes in cascade_chain than to fix it in cascade_chain and later
backport the changes to the main branch.