Fixed GitHub workflow files to use the correct Python3 package
This is unrelated to the logger changes in this branch, but it needs to
be fixed to get the "build check" to pass on GitHub. GitHub recently
updated their build servers to a later version of Ubuntu, which does not
have a "python3.10" package - it has one named "python3.12" instead. The
better way to ensure the latest python is installed in these build
scripts is to install the "python3" package, which always points to the
current Python 3 package in Ubuntu.