Improved documentation in service.hpp, fixed compiler warnings
The comments on the main API functions in service.hpp were missing some
details and were sometimes out-of-date with respect to the code. Also,
a newer g++ (version 13.3) emitted some warnings about existing code
that needed to be updated (e.g. to remove unnecessary std::moves).
Finally, the string constants in service.hpp should be declared as
constexpr strings, not macros, to match Derecho's usage of these types
of constants for Conf strings.