Added debug logging to ServiceClient
The main API functions in ServiceClient should have "entered" and
"leaving" debug log statements, like the ones on the server side in
CascadeStore, so that I can debug client problems more easily. It's hard
to know what's going wrong if I don't know if/when the client sent a
request.