Added debug output to ShaHashUDL
It can't use the Logger framework used by the rest of Cascade, but
sometimes it would be nice to have it print out what it is doing. It
will just have to use stdout, but I added a config option to disable the
output so it can be turned off in production mode.