Partial bug fix for signature notification handler matching
While testing Cascade CBDC with CascadeChain, we noticed a bug: the
object_pool_pathname field SignatureCascadeStore puts in the
notification message is not necessarily equal to the object pool name. I
used get_pathname(key), but that returns the entire path prefix before
the key, and the object pool name might be a shorter prefix (e.g. in
CBDC, the object pool is /cbdc/log/ but keys are of the form
/cbdc/log/r/m/WID_1997). Nothing on the server side checks if the
object_pool_pathname field matches an object pool, but the client side
notification handler (SubgroupNotificationHandler) only allows you to
register notification callbacks that match object pools exactly, and
only calls callbacks for notification messages whose
object_pool_pathname exactly matches a registered object pool name.

Existing users of the Cascacde notification messages (mostly Cascade
DDS) call notify from a UDL, where the UDL's prefix_length argument can
be used to determine what part of the path is the object pool name.
Within SignatureCascadeStore, though, there's no way to determine what
object pool a key is in, other than asking CascadeMetadataService
(through ServiceClient's find_object_pool), which would incur an extra
network roundtrip to a CascadeMetadataService node.

For now, I can work around this on the client side by making the
SubgroupNotificationHandler for signature messages check each sub-path
of the object_pool_pathname field to see if it matches an installed
object pool callback. This is just a few extra string operations and
hashtable lookups, so hopefully it's not a lot of overhead compared to
querying CascadeMetadataService.

A better solution would be to either find a way to send the signature
notifications from the Off-Critical-Path engine that calls UDLs, which
already knows which object pool prefixes match each key, or remove the
client's assumption that object_pool_pathname is an exact match for an
object pool and use a prefix tree to match notifications to handlers.
For now, though, we need to get Cascade CBDC working as soon as
possible.