Made CascadeChain sample configs use smaller port numbers
As I noted while fixing bugs in Derecho (PR #288), the default port
numbers we've been using in derecho.cfg are too large and can easily
overlap with the Linux ephemeral port range. Any port higher than 1024
is not reserved for standard services, and as long as we don't use
common "local webserver" ports like 8888 or 8080 it should be fine to
use ports under 10000 for CascadeChain.